+++ public/app/features/dashboard-scene/scene/layout-auto-grid/AutoGridLayoutManager.tsx
