+++ public/app/features/dashboard-scene/scene/layouts-shared/layoutRegistry.ts
@@ -1,7 +1,7 @@
 import { Registry } from '@grafana/data';
 
+import { AutoGridLayoutManager } from '../layout-auto-grid/AutoGridLayoutManager';
 import { DefaultGridLayoutManager } from '../layout-default/DefaultGridLayoutManager';
-import { AutoGridLayoutManager } from '../layout-responsive-grid/AutoGridLayoutManager';
 import { RowsLayoutManager } from '../layout-rows/RowsLayoutManager';
 import { TabsLayoutManager } from '../layout-tabs/TabsLayoutManager';
 import { LayoutRegistryItem } from '../types/LayoutRegistryItem';
